fix bug

+++ checkattempts.php
@@ -50,12 +50,12 @@
             } catch (\Throwable $th) {
                 $create_table_query = "CREATE TABLE $sql_table(
                     ATTEMPT_ID INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
-                    STUDENT_ID INT NOT NULL,
+                    STUDENT_ID BIGINT NOT NULL,
                     FIRST_NAME VARCHAR(30) NOT NULL,
                     LAST_NAME VARCHAR (30) NOT NULL,
-                    SCORE INT NOT NULL,
+                    SCORE TINYINT NOT NULL,
                     CREATED_AT DATETIME NOT NULL,
-                    ATTEMPT_NUMBER INT NOT NULL,
+                    ATTEMPT_NUMBER TINYINT NOT NULL,
                     FOREIGN KEY(STUDENT_ID) REFERENCES students(STUDENT_ID)
                 );";
                 $result = @mysqli_query($conn, $create_table_query);
